+++ src/ng-module.js
@@ -1,1 +1,1 @@
-module.exports = require('angular').module('angular-popup', []);
+module.exports = require('angular').module('angular-popups', []);