+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 ## 0.0.1
 
-1. 项目正式命名为 `angular-popup`
-2. 提供无 CSS 依赖的版本 
+1. 提供无 CSS 依赖的版本 
 
 ## 0.0.1-beta9
 
@@ -38,7 +37,7 @@
 
 ## 0.0.1-beta2
 
-1. 模块名由 `popups` 改成 `angular-popups`
+1. 模块名由 `popups` 改成 `angular-popupss`
 2. 修复按按 ESC 关闭浮层，浮层没有恢复焦点的问题
 3. 修复 `align` 无效的问题
 4. 优化 `bubble` 指令样式